Refactor SpringAnim parameters and update time input type
- Update `calculateSpringOscillation` to accept a `double` time input.
- Standardize argument string conversion in `main.cpp` by using `std::stod`.

+++ extension/bce_anim_engine/include/calculate.h
@@ -28,5 +28,5 @@ public:
 			
 	std::string getName() const;
 	std::string getParams() const;
-	double calculateSpringOscillation(int _t) const;
+	double calculateSpringOscillation(double _t) const;
 };
